Keras Project
The goal of this Project is to explore the deep learning framework known as keras. With Conv2D (for creating a convolution layer), MaxPooling2D (for a max pooling layer), AveragePooling2D (for average pooling, as an alternative to max pooling), Dense (for creating hidden and output layers), Optimizers (for determining a weight training algorithm). 
Some of the parameters that you can modify are:
• Max vs average pooling
• Number of hidden (Dense) layers, number of neurons in each layer, and activation function for each layer (last Dense line is output layers and must have neurons equal to 10 and sigmoid activation)
• Learning rate
• Optimizer (for weight training; see documentation in link above)
• Number of epochs (in “fit” function)